Refactor: Extract domain-specific functions into separate utility modules
- Created utils/data_cleaners.py for data cleaning functions (clean_chems, clean_equipment)
- Created utils/industry_analysis.py for industrial data analysis (analyze_grouped)
- Created utils/firefighter_analysis.py for firefighter survey analysis
- Reduced Read_excels_as_one.py from 573 to 319 lines (44% reduction)
- Removed unused imports (os, numpy)
- Fixed type hints and improved code formatting
- Added comprehensive docstrings to analyze_ff_survey_files
